Update README.md
In this lesson students are expected to be able to understand something or other values of a word or word arrangement that becomes a sentence. In doing the preparation of sentences of ordinary students can embedd well as well, we used to call it with Mantra values. But this method of giving value to the initial kelaas is only a lesson in word meaning only and we usually for students the best is that they have high Karmic values.